Added BigDecimal currency handling for greater accuracy. Tested new Maxi Savings interest rate. Removed deprecated functions.

+++ src/test/java/com/abc/TransactionTest.java
@@ -2,12 +2,14 @@ package com.abc;
 
 import org.junit.Test;
 
+import java.math.BigDecimal;
+
 import static org.junit.Assert.assertTrue;
 
 public class TransactionTest {
     @Test
     public void transaction() {
-        Transaction t = new Transaction(5);
+        Transaction t = new Transaction(BigDecimal.valueOf(5));
         assert t instanceof Transaction;
     }
 }
